routes, controllers, and models done with extra authorization for admin actions
booking, hotel, package, room, wishlist pages done

+++ backend/controllers/hotelController.js
@@ -2,55 +2,87 @@
 
 import mongoose from "mongoose";
 
-import  Hotel from "../Models/hotelModel.js";
+import Hotel from "../Models/hotelModel.js";
+import Room from "../Models/roomModel.js";
+import authorizationChecker from "../middleware/auth.js"
+import { authorize } from "../middleware/auth.js";
 
 
 
 export const getAllHotels = async (req, res) => {
-let location;
-if(!req.query.city){
-location={}
-}
-else{
-    location =req.query.city
-}
-try {
-    const auth = await authorizationChecker(req)
-    
-    if (auth === "A") {
-        return res.status(401).json({ message: "token reqired" })
+    let location;
+    if (!req.query.city) {
+        location = {}
     }
-    else if (auth === "C") {
-        return res.status(401).json({ message: "not auth" })
+    else {
+        location = req.query.city
     }
-    const hotels =await Hotel.find({location})
-    res.status(200).json({hotels}) 
-} catch (error) {
-    res.status(500).json({msg:error.message})
-}
-
-
-
-
-};
-export const getHotel = async (req, res) => {
     try {
         const auth = await authorizationChecker(req)
-    
+
         if (auth === "A") {
             return res.status(401).json({ message: "token reqired" })
         }
         else if (auth === "C") {
             return res.status(401).json({ message: "not auth" })
         }
-        const id =req.params.id
-        const hotel =await Hotel.find({_id:id})
-        if(!hotel){
-            return res.status(404).json({msg:"No Hotel "})
-        }
-        res.status(200).json({data:hotel}) 
+        const hotels = await Hotel.find({ location })
+        res.status(200).json({ hotels })
     } catch (error) {
-        res.status(500).json({msg:error.message})
+        res.status(500).json({ msg: error.message })
+    }
+}
+
+export const getHotelRoom = async (req, res) => {
+
+  
+    try {
+        const auth = await authorizationChecker(req)
+
+        if (auth === "A") {
+            return res.status(401).json({ message: "token reqired" })
+        }
+        else if (auth === "C") {
+            return res.status(401).json({ message: "not auth" })
+        }
+        const id = req.params.id
+        let rooms;
+        if (!req.query.taken) {
+            rooms = await Room.find({ hotel: id })
+        }
+        else {
+            rooms = await Room.find({ hotel: id, taken: req.query.taken })
+        }
+        if (!rooms) {
+            return res.status(404).json({ msg: "No Hotel " })
+        }
+        res.status(200).json({ data: rooms })
+    } catch (error) {
+        res.status(500).json({ msg: error.message })
+    }
+
+
+}
+
+
+export const getHotel = async (req, res) => {
+    try {
+        const auth = await authorizationChecker(req)
+
+        if (auth === "A") {
+            return res.status(401).json({ message: "token reqired" })
+        }
+        else if (auth === "C") {
+            return res.status(401).json({ message: "not auth" })
+        }
+        const id = req.params.id
+        const hotel = await Hotel.find({ _id: id })
+        if (!hotel) {
+            return res.status(404).json({ msg: "No Hotel " })
+        }
+        res.status(200).json({ data: hotel })
+    } catch (error) {
+        res.status(500).json({ msg: error.message })
     }
 
 
@@ -64,13 +96,15 @@ export const addHotel = async (req, res) => {
         else if (auth === "C") {
             return res.status(401).json({ message: "not auth" })
         }
-        const hotel=await Hotel.create(req.body)
-        res.status(200).json({success:true,data:hotel})
+        console.log(auth)
+        authorize(res,auth,"admin")
+        const hotel = await Hotel.create(req.body)
+        res.status(200).json({ success: true, data: hotel })
     } catch (error) {
-        res.status(500).json({msg:error.message})
+        res.status(500).json({ msg: error.message })
     }
 };
-export const updateHotel = async (req, res ) => {
+export const updateHotel = async (req, res) => {
     try {
         const auth = await authorizationChecker(req)
         if (auth === "A") {
@@ -79,32 +113,34 @@ export const updateHotel = async (req, res ) => {
         else if (auth === "C") {
             return res.status(401).json({ message: "not auth" })
         }
-        const hotel=await Hotel.findOneAndUpdate({_id:req.params.id,},req.body,{new:true})
-        if(!hotel){
-            return res.status(404).json({msg:"No Hotel "})
+        authorize(res,auth,"admin")
+        const hotel = await Hotel.findOneAndUpdate({ _id: req.params.id, }, req.body, { new: true })
+        if (!hotel) {
+            return res.status(404).json({ msg: "No Hotel " })
         }
-        res.status(200).json({success:true,data:hotel})
+        res.status(200).json({ success: true, data: hotel })
     } catch (error) {
-        res.status(500).json({msg:error.message})
+        res.status(500).json({ msg: error.message })
     }
 
 };
 export const deleteHotel = async (req, res) => {
     try {
         const auth = await authorizationChecker(req)
-        
-    
+
+
         if (auth === "A") {
             return res.status(401).json({ message: "token reqired" })
         }
         else if (auth === "C") {
             return res.status(401).json({ message: "not auth" })
         }
-        const hotel=await Hotel.findOneAndDelete({_id:req.params.id})
-        res.status(200).json({data:hotel})
+        authorize(res,auth,"admin")
+        const hotel = await Hotel.findOneAndDelete({ _id: req.params.id })
+        res.status(200).json({ data: hotel })
     }
-        catch(error){
-            res.status(500).json({msg:error.message}) 
-        }
+    catch (error) {
+        res.status(500).json({ msg: error.message })
+    }
 
 };
